@@ -60,7 +60,7 @@ dependencies:
   flutter_app_lock: ^4.0.0-dev.0
   sqlite3_flutter_libs: ^0.5.13
   sqflite_common_ffi: ^2.2.3
-  sentry_flutter: ^7.5.2
+  sentry_flutter: ^7.10.1
   flutter_dotenv: ^5.0.2
   contextmenu: ^3.0.0
   window_size:
